refactor: Simply moved movie command and tests to own package
